finished and reached the same state as the example, but not really happy with the way the asyncronous stuff works

+++ mathfun.c
@@ -1,8 +1,8 @@
-int fibo(int index){
-    int cur = 0;
-    int next = 1;
-    for (int i = 0; i < index; i++){
-        int temp = cur;
+unsigned long long fibo(unsigned long index){
+    unsigned long cur = 0;
+    unsigned long next = 1;
+    for (unsigned long i = 0; i < index; i++){
+        unsigned long temp = cur;
         cur = next;
         next += temp;
     }
