+++ mathfun.h
@@ -1,6 +1,6 @@
 #ifndef FIBO_H
 #define FIBO_H
 
-extern int fibo(int index);
+extern unsigned long long fibo(unsigned long index);
 
 #endif
